fix: downgrade Android DFU library to 2.6.0 for V1 (#3)
This is the newest version that's working with micro:bit V1 (V2 is OK with 2.9 and 2.6) and matches what's used in the current micro:bit app.
This broke in 2.7 and is still broken in 2.10.1.
Just downgrading as a workaround for now.
There are some interesting changes to review including some that might be able to be reverted by setting options.